requirements.txt
boto3==1.20.19
botocore==1.23.19
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
zipp==3.6.0
